Release ver_2.2.9

+++ os/kernel/include/chqueues.h
@@ -38,16 +38,16 @@
 
 #if CH_USE_QUEUES || defined(__DOXYGEN__)
 
-/**
- * @name    Queue functions returned status value
- * @{
- */
-#define Q_OK            RDY_OK      /**< @brief Operation successful.       */
-#define Q_TIMEOUT       RDY_TIMEOUT /**< @brief Timeout condition.          */
-#define Q_RESET         RDY_RESET   /**< @brief Queue has been reset.       */
-#define Q_EMPTY         -3          /**< @brief Queue empty.                */
-#define Q_FULL          -4          /**< @brief Queue full,                 */
-/** @} */
+/** @brief Returned by the queue functions if the operation is successful.*/
+#define Q_OK            RDY_OK
+/** @brief Returned by the queue functions if a timeout occurs.*/
+#define Q_TIMEOUT       RDY_TIMEOUT
+/** @brief Returned by the queue functions if the queue has been reset.*/
+#define Q_RESET         RDY_RESET
+/** @brief Returned by the queue functions if the queue is empty.*/
+#define Q_EMPTY         -3
+/** @brief Returned by the queue functions if the queue is full.*/
+#define Q_FULL          -4
 
 /**
  * @brief   Type of a generic I/O queue structure.
@@ -78,10 +78,6 @@ struct GenericQueue {
 };
 
 /**
- * @name    Macro Functions
- * @{
- */
-/**
  * @brief   Returns the queue's buffer size.
  *
  * @param[in] qp        pointer to a @p GenericQueue structure.
@@ -101,8 +97,7 @@ struct GenericQueue {
  *
  * @iclass
  */
-#define chQSpaceI(qp) ((qp)->q_counter)
-/** @} */
+#define chQSpaceI(qp) ((size_t)((qp)->q_counter))
 
 /**
  * @extends GenericQueue
@@ -117,10 +112,6 @@ struct GenericQueue {
  */
 typedef GenericQueue InputQueue;
 
-/**
- * @name    Macro Functions
- * @{
- */
 /**
  * @brief   Returns the filled space into an input queue.
  *
@@ -181,7 +172,6 @@ typedef GenericQueue InputQueue;
  * @api
  */
 #define chIQGet(iqp) chIQGetTimeout(iqp, TIME_INFINITE)
-/** @} */
 
 /**
  * @brief   Data part of a static input queue initializer.
@@ -229,19 +219,15 @@ typedef GenericQueue InputQueue;
  */
 typedef GenericQueue OutputQueue;
 
-/**
- * @name    Macro Functions
- * @{
- */
-/**
- * @brief   Returns the filled space into an output queue.
- *
- * @param[in] oqp       pointer to an @p OutputQueue structure
- * @return              The number of full bytes in the queue.
- * @retval 0            if the queue is empty.
- *
- * @iclass
- */
+ /**
+  * @brief   Returns the filled space into an output queue.
+  *
+  * @param[in] oqp       pointer to an @p OutputQueue structure
+  * @return              The number of full bytes in the queue.
+  * @retval 0            if the queue is empty.
+  *
+  * @iclass
+  */
 #define chOQGetFullI(oqp) (chQSizeI(oqp) - chQSpaceI(oqp))
 
 /**
@@ -295,7 +281,6 @@ typedef GenericQueue OutputQueue;
  * @api
  */
 #define chOQPut(oqp, b) chOQPutTimeout(oqp, b, TIME_INFINITE)
- /** @} */
 
 /**
  * @brief   Data part of a static output queue initializer.
